+++ demo/src/test/java/com/evernote/android/state/test/TestBundling.java
@@ -26,7 +26,7 @@ import org.robolectric.annotation.Config;
 
 import java.util.ArrayList;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Java6Assertions.assertThat;
 
 /**
  * @author rwondratschek