+++ library-lint/src/test/resources/java/InvalidActivityNoRestoreKt.kt
@@ -1,0 +1,29 @@
+/* *****************************************************************************
+ * Copyright (c) 2017 Evernote Corporation.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *    Ralf Wondratschek - initial version
+ *******************************************************************************/
+package com.evernote.android.state.lint
+
+import android.app.Activity
+import android.os.Bundle
+import com.evernote.android.state.StateSaver
+
+/**
+ * @author rwondratschek
+ */
+class InvalidActivityNoRestoreKt : Activity() {
+    protected override fun onCreate(savedInstanceState: Bundle) {
+        super.onCreate(savedInstanceState)
+    }
+
+    protected override fun onSaveInstanceState(outState: Bundle) {
+        super.onSaveInstanceState(outState)
+        StateSaver.saveInstanceState(this, outState)
+    }
+}