+++ library-lint/src/test/resources/java/ValidActivity.java
@@ -1,0 +1,32 @@
+/* *****************************************************************************
+ * Copyright (c) 2017 Evernote Corporation.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *    Ralf Wondratschek - initial version
+ *******************************************************************************/
+package com.evernote.android.state.lint;
+
+import android.app.Activity;
+import android.os.Bundle;
+import com.evernote.android.state.StateSaver;
+
+/**
+ * @author rwondratschek
+ */
+public class ValidActivity extends Activity {
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        StateSaver.restoreInstanceState(this, savedInstanceState);
+    }
+
+    @Override
+    protected void onSaveInstanceState(Bundle outState) {
+        super.onSaveInstanceState(outState);
+        StateSaver.saveInstanceState(this, outState);
+    }
+}