+++ processor/src/main/java/com/evernote/android/state/StateProcessor.java
@@ -116,6 +116,7 @@ public class StateProcessor extends AbstractProcessor {
     };
 
     private static final String STATE_CLASS_NAME = State.class.getName();
+    private static final String STATE_REFLECTION_CLASS_NAME = StateReflection.class.getName();
     private static final String OBJECT_CLASS_NAME = Object.class.getName();
     private static final String PARCELABLE_CLASS_NAME = Parcelable.class.getName();
     private static final String SERIALIZABLE_CLASS_NAME = Serializable.class.getName();
@@ -151,7 +152,10 @@ public class StateProcessor extends AbstractProcessor {
 
     @Override
     public Set<String> getSupportedAnnotationTypes() {
-        return Collections.singleton(State.class.getName());
+        Set<String> annotations = new HashSet<>();
+        annotations.add(State.class.getName());
+        annotations.add(StateReflection.class.getName());
+        return Collections.unmodifiableSet(annotations);
     }
 
     @Override
@@ -161,7 +165,10 @@ public class StateProcessor extends AbstractProcessor {
 
     @Override
     public boolean process(Set<? extends TypeElement> annotations, RoundEnvironment env) {
-        final Set<? extends Element> annotatedFields = env.getElementsAnnotatedWith(State.class);
+        final Set<Element> annotatedFields = new HashSet<>();
+        annotatedFields.addAll(env.getElementsAnnotatedWith(State.class));
+        annotatedFields.addAll(env.getElementsAnnotatedWith(StateReflection.class));
+
         final Map<Element, BundlerWrapper> bundlers = new HashMap<>();
 
         for (Element field : annotatedFields) {
@@ -309,9 +316,13 @@ public class StateProcessor extends AbstractProcessor {
                         break;
 
                     case BOOLEAN_PROPERTY:
+                    case BOOLEAN_PROPERTY_KOTLIN:
                     case PROPERTY:
-                        saveMethodBuilder.addStatement("HELPER.put$N(state, $S, target.$N$N())", mapping, fieldName,
-                                fieldType == FieldType.BOOLEAN_PROPERTY ? "is" : "get", fieldName);
+                        if (fieldType == FieldType.BOOLEAN_PROPERTY || fieldType == FieldType.BOOLEAN_PROPERTY_KOTLIN) {
+                            saveMethodBuilder.addStatement("HELPER.put$N(state, $S, target.$N$N())", mapping, fieldName, "is", fieldName);
+                        } else {
+                            saveMethodBuilder.addStatement("HELPER.put$N(state, $S, target.$N$N())", mapping, fieldName, "get", fieldName);
+                        }
 
                         if (bundler != null) {
                             restoreMethodBuilder = restoreMethodBuilder.addStatement("target.set$N(HELPER.<$T>get$N(state, $S))", fieldName,
@@ -518,10 +529,22 @@ public class StateProcessor extends AbstractProcessor {
         String isGetterName = "is" + fieldName;
         String setterName = "set" + fieldName;
 
+        String isGetterNameKotlin = null;
+        String setterNameKotlin = null;
+
+        if (fieldName.length() > 2 && fieldName.startsWith("Is") && Character.isUpperCase(fieldName.charAt(2))) {
+            // cut off the "is"
+            String substring = fieldName.substring(2);
+            isGetterNameKotlin = "is" + substring;
+            setterNameKotlin = "set" + substring;
+        }
+
+
         List<? extends Element> elements = field.getEnclosingElement().getEnclosedElements();
         boolean hasGetter = false;
         boolean hasIsGetter = false;
         boolean hasSetter = false;
+        boolean isKotlinBooleanProperty = false;
 
         for (Element element : elements) {
             if (element.getKind() != ElementKind.METHOD) {
@@ -540,8 +563,20 @@ public class StateProcessor extends AbstractProcessor {
                     break;
                 }
             }
-
+            if (!hasGetter && isGetterNameKotlin != null && isGetterNameKotlin.equals(elementName) && !element.getModifiers().contains(Modifier.PRIVATE)) {
+                hasIsGetter = true;
+                isKotlinBooleanProperty = true;
+                if (hasSetter) {
+                    break;
+                }
+            }
             if (!hasSetter && setterName.equals(elementName) && !element.getModifiers().contains(Modifier.PRIVATE)) {
+                hasSetter = true;
+                if (hasGetter) {
+                    break;
+                }
+            }
+            if (!hasSetter && setterNameKotlin != null && setterNameKotlin.equals(elementName) && !element.getModifiers().contains(Modifier.PRIVATE)) {
                 hasSetter = true;
                 if (hasGetter) {
                     break;
@@ -549,7 +584,9 @@ public class StateProcessor extends AbstractProcessor {
             }
         }
 
-        if (hasIsGetter && hasSetter) {
+        if (isKotlinBooleanProperty && hasSetter) {
+            return FieldType.BOOLEAN_PROPERTY_KOTLIN;
+        } else if (hasIsGetter && hasSetter) {
             return FieldType.BOOLEAN_PROPERTY;
         } else if (hasGetter && hasSetter) {
             return FieldType.PROPERTY;
@@ -559,7 +596,7 @@ public class StateProcessor extends AbstractProcessor {
     }
 
     private enum FieldType {
-        FIELD, FIELD_REFLECTION, PROPERTY, BOOLEAN_PROPERTY, NOT_SUPPORTED;
+        FIELD, FIELD_REFLECTION, PROPERTY, BOOLEAN_PROPERTY, BOOLEAN_PROPERTY_KOTLIN, NOT_SUPPORTED;
 
         public String getFieldName(Element field) {
             switch (this) {
@@ -569,6 +606,8 @@ public class StateProcessor extends AbstractProcessor {
                 case PROPERTY:
                 case BOOLEAN_PROPERTY:
                     return getPropertyFieldName(field);
+                case BOOLEAN_PROPERTY_KOTLIN:
+                    return field.getSimpleName().toString().substring(2);
                 case NOT_SUPPORTED:
                 default:
                     return null;
@@ -578,7 +617,7 @@ public class StateProcessor extends AbstractProcessor {
 
     private BundlerWrapper getBundlerWrapper(Element field) {
         for (AnnotationMirror annotationMirror : field.getAnnotationMirrors()) {
-            if (!STATE_CLASS_NAME.equals(annotationMirror.getAnnotationType().toString())) {
+            if (!isStateAnnotation(annotationMirror)) {
                 continue;
             }
 
@@ -633,19 +672,8 @@ public class StateProcessor extends AbstractProcessor {
 
     private boolean useReflection(Element field) {
         for (AnnotationMirror annotationMirror : field.getAnnotationMirrors()) {
-            if (!STATE_CLASS_NAME.equals(annotationMirror.getAnnotationType().toString())) {
-                continue;
-            }
-
-            Map<? extends ExecutableElement, ? extends AnnotationValue> elementValues = annotationMirror.getElementValues();
-            for (ExecutableElement executableElement : elementValues.keySet()) {
-                if ("reflection".equals(executableElement.getSimpleName().toString())) {
-                    Object value = elementValues.get(executableElement).getValue(); // bundler class
-                    if (value == null) {
-                        continue;
-                    }
-                    return "true".equals(value.toString());
-                }
+            if (STATE_REFLECTION_CLASS_NAME.equals(annotationMirror.getAnnotationType().toString())) {
+                return true;
             }
         }
         return false;
@@ -811,5 +839,10 @@ public class StateProcessor extends AbstractProcessor {
             }
         }
         return mLicenseHeader;
+    }
+
+    private boolean isStateAnnotation(AnnotationMirror annotationMirror) {
+        String string = annotationMirror.getAnnotationType().toString();
+        return STATE_CLASS_NAME.equals(string) || STATE_REFLECTION_CLASS_NAME.equals(string);
     }
 }